Move host specific funcs back to HostBase.
