+++ networkml/featurizers/funcs/host.py
@@ -3,7 +3,10 @@ import ipaddress
 import statistics
 from numpy import percentile
 from networkml.featurizers.features import Features
+import netaddr
 
+
+MAC_BCAST = netaddr.EUI('FF-FF-FF-FF-FF-FF')
 ETH_TYPE_ARP = 0x806
 ETH_TYPE_IP = 0x800
 ETH_TYPE_IPV6 = 0x86DD
@@ -34,6 +37,56 @@ class HostBase:
     WK_NONPRIV_PROTO_PORTS = frozenset(
         [1900, 2375, 2376, 5222, 5349, 5353, 5354, 5349, 5357, 6653])
     WK_PROTOS = frozenset(('tcp', 'udp', 'icmp', 'icmpv6', 'arp', 'other'))
+
+    @staticmethod
+    def _is_unicast(mac):
+        mac_val = netaddr.EUI(mac)
+        if mac_val == MAC_BCAST or netaddr.EUI(mac_val).packed[0] & 1:
+            return False
+        return True
+
+    def _tshark_input_mac(self, rows):
+        '''Infer MAC address of host connected to a port, from pcap.
+
+           Pick the host that most often occurs as both source and destination.
+        '''
+        eth_srcs = Counter()
+        eth_dsts = Counter()
+        all_eths = Counter()
+
+        for row in rows:
+            eth_src = row.get('eth.src', None)
+            if eth_src and self._is_unicast(eth_src):
+                eth_srcs[eth_src] += 1
+                all_eths[eth_src] += 1
+            eth_dst = row.get('eth.dst', None)
+            if eth_dst and self._is_unicast(eth_dst):
+                eth_dsts[eth_dst] += 1
+                all_eths[eth_dst] += 1
+
+        common_eth = set(eth_srcs).union(eth_dsts)
+        if len(common_eth) > 1:
+            common_count = [(eth, all_eths[eth]) for eth in common_eth]
+            max_eth = sorted(common_count, key=lambda x: x[1])[-1][0]
+            return (max_eth, set(all_eths.keys()))
+        return (None, None)
+
+    def _select_mac_direction(self, rows, output=True):
+        '''Return filter expression selecting input or output rows.'''
+        src_mac, all_macs = self._tshark_input_mac(rows)
+        # Select all if can't infer direction.
+        if src_mac is None or len(all_macs) == 1:
+            return rows
+        # We have to return rows for all hosts whether in or out
+        # to ensure columns are always present.
+        placeholder_rows = [{'eth.src': eth_src} for eth_src in all_macs]
+        if output:
+            # Select all rows where traffic originated by inferred source MAC
+            return filter(lambda row: (
+                row.get('eth.src', None) == src_mac or len(row) == 1), rows + placeholder_rows)
+        # Select all rows where traffic not originated by inferred source MAC.
+        return filter(lambda row: (
+            row.get('eth.src', None) != src_mac or len(row) == 1), rows + placeholder_rows)
 
     def _pyshark_ipversions(self, rows):
         ipversions = set()
